refactor(ui): extract PcInfoTabMixin (Tab 1 — the largest)
Move build_ui_page_1 plus all of its sibling handlers/setters out of
L5RMain into l5r/ui/tabs/pc_info.py:

* build_ui_page_1 (identity card, rings/attributes grid, honor/glory/
  status/taint/infamy flags, void points, initiative, armor TN, wounds)
* signal handlers: on_trait_increase, on_void_increase,
  on_pc_name_change, on_pers_info_change, on_flag_points_change,
  on_flag_rank_change, on_void_points_change
* model -> UI setters: set_clan, set_family, set_school,
  set_void_points, set_flag and the five set_{honor,glory,status,
  taint,infamy} wrappers

main.py: 1674 -> 1204 lines.
